requirements.txt
flask
transformers
torch
pillow
googletrans==4.0.0rc1